リリース対応 パッケージ名の書き換え
