+++ app/src/main/java/com/channaru/doubleroulette/ArrowViewFragment.kt
@@ -1,4 +1,4 @@
-package com.example.doubleroulette
+package com.channaru.doubleroulette
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
