+++ app/src/main/java/com/channaru/doubleroulette/DoubleRouletteApplication.kt
@@ -1,4 +1,4 @@
-package com.example.doubleroulette
+package com.channaru.doubleroulette
 
 import android.app.Application
 import io.realm.Realm
