+++ app/src/main/java/com/channaru/doubleroulette/OuterRouletteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.doubleroulette
+package com.channaru.doubleroulette
 
 import android.os.Build
 import android.os.Bundle
@@ -10,7 +10,7 @@ import androidx.annotation.RequiresApi
 import io.realm.Realm
 import io.realm.kotlin.where
 
-class InnerRouletteFragment : Fragment() {
+class OuterRouletteFragment : Fragment() {
 
     private lateinit var realm: Realm
 
@@ -24,10 +24,10 @@ class InnerRouletteFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val innerRouletteData = realm.where<DoubleRouletteModel>().equalTo("isInner", true).findAll()
-        innerRouletteData?.let {
+        val outerRouletteData = realm.where<DoubleRouletteModel>().equalTo("isInner", false).findAll()
+        outerRouletteData?.let {
             if (it.count() > 0) {
-                return PieChartView(this.requireContext(), it, true)
+                return PieChartView(this.requireContext(), it, false)
             }
         }
         return null
