+++ app/src/main/java/com/channaru/doubleroulette/PieChartView.kt
@@ -1,4 +1,4 @@
-package com.example.doubleroulette
+package com.channaru.doubleroulette
 
 import android.content.Context
 import android.graphics.*
