+++ app/src/main/res/layout/activity_roulette.xml
@@ -60,7 +60,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/outerRouletteFragmentView"
-        android:name="com.example.doubleroulette.OuterRouletteFragment"
+        android:name="com.channaru.doubleroulette.OuterRouletteFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/adAreaView"
@@ -73,7 +73,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/innerRouletteFragmentView"
-        android:name="com.example.doubleroulette.InnerRouletteFragment"
+        android:name="com.channaru.doubleroulette.InnerRouletteFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/adAreaView"
@@ -100,7 +100,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/arrowFragmentView"
-        android:name="com.example.doubleroulette.ArrowViewFragment"
+        android:name="com.channaru.doubleroulette.ArrowViewFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/adAreaView"
